Make several declarations public
Rearrange some typedefs so that a user need only read the public
wavltree.h rather than also needing to read the wavltree_priv.h in
order to know how to integrate a tree into their code.
