+++ wavltree.h
@@ -1,7 +1,40 @@
 #pragma once
 
+#include <stdint.h>
+
+/**
+ * A WAVL-tree node. Embed this in your own structure. All members of this structure
+ * are private.
+ */
 struct wavl_tree_node;
+
+/**
+ * A WAVL tree. This structure contains all the state needed to maintain a wavl
+ * tree. All members of this structure are private, and should not be inspected or
+ * modified by users.
+ */
 struct wavl_tree;
+
+/**
+ * A normal WAVL tree result
+ */
+typedef uint32_t wavl_result_t;
+
+/**
+ * Ordering function to compare a node to another node.
+ */
+typedef wavl_result_t (*wavl_node_to_node_compare_func_t)(struct wavl_tree *tree,
+                                                          struct wavl_tree_node *lhs,
+                                                          struct wavl_tree_node *rhs,
+                                                          int *pdir);
+
+/**
+ * Ordering function to compare a node to a key.
+ */
+typedef wavl_result_t (*wavl_key_to_node_compare_func_t)(struct wavl_tree *tree,
+                                                         void *key_lhs,
+                                                         struct wavl_tree_node *rhs,
+                                                         int *pdir);
 
 #define __WAVL_INCLUDING_WAVL_PRIV_H__
 #include "wavltree_priv.h"
